bump to expo 52
