+++ examples/fishjam-chat/package.json
@@ -19,12 +19,12 @@
     "@react-navigation/bottom-tabs": "^6.6.1",
     "@react-navigation/native": "^6.1.18",
     "@react-navigation/native-stack": "^6.11.0",
-    "expo": "^51.0.32",
+    "expo": "^52.0.0",
     "expo-build-properties": "~0.12.5",
     "expo-device": "^6.0.2",
     "expo-status-bar": "~1.12.1",
-    "react": "^18.2.0",
-    "react-native": "^0.75.2",
+    "react": "18.3.1",
+    "react-native": "0.76.1",
     "react-native-gesture-handler": "~2.20.2",
     "react-native-reanimated": "^3.16.1",
     "react-native-safe-area-context": "4.14.0",
