+++ packages/react-native-client/plugin/src/withFishjamIos.ts
@@ -16,7 +16,7 @@ target '${SBE_TARGET_NAME}' do
   pod 'FishjamCloudClient/Broadcast'
 end`;
 const TARGETED_DEVICE_FAMILY = `"1,2"`;
-const IPHONEOS_DEPLOYMENT_TARGET = '13.4';
+const IPHONEOS_DEPLOYMENT_TARGET = '15.1';
 const GROUP_IDENTIFIER_TEMPLATE_REGEX = /{{GROUP_IDENTIFIER}}/gm;
 const BUNDLE_IDENTIFIER_TEMPLATE_REGEX = /{{BUNDLE_IDENTIFIER}}/gm;
