+++ packages/react-native-client/src/hooks/useAudioSettings.ts
@@ -1,8 +1,9 @@
 import { useCallback, useEffect } from 'react';
 import { Platform } from 'react-native';
 
-import RNFishjamClientModule from '../RNFishjamClientModule';
-import { ReceivableEvents } from './useFishjamEvent';
+import RNFishjamClientModule, {
+  ReceivableEvents,
+} from '../RNFishjamClientModule';
 import { useFishjamEventState } from './useFishjamEventState';
 
 export type AudioOutputDeviceType =
